Refactor ProductAdapter
