+++ apps/rappi-backend-m2/src/products/infrastructure/adapters/product.adapter.ts
@@ -1,26 +1,21 @@
 import { Inject, Injectable } from "@nestjs/common";
+
 import { PRODUCT_REPOSITORY } from "../constants/product-repository.constants";
 import { IProductAdapter } from "../../domain/interfaces/IProductAdapter";
-import { ProductOfItem } from "../../../order/domain/entities/product-of-item.entity";
 import { IProductRepository } from '../../domain/interfaces/IProductRepository';
-import { Product } from "../../domain/entities/product.entity";
+import { ProductOfItem } from "../../domain/dtos/product-of-item.entity";
 
 @Injectable()
 export class ProductAdapter implements IProductAdapter {
-    constructor(
-        @Inject(PRODUCT_REPOSITORY)
-        private readonly productRepository: IProductRepository
-    ) { }
+  constructor(
+    @Inject(PRODUCT_REPOSITORY)
+    private readonly productRepository: IProductRepository
+  ) {}
 
-    async getProductById(id: string): Promise<ProductOfItem | null> {
-        const entity = await this.productRepository.findById(id);
-        if(!entity) return null;
-        return fromEntity(entity);
-    }
+  async getProductById(id: string): Promise<ProductOfItem | null> {
+    const entity = await this.productRepository.findById(id);
+    if (!entity) return null;
+    
+    return ProductOfItem.fromEntity(entity);
+  }
 }
-
-
-
-    function fromEntity(entity: Product): ProductOfItem {
-        return new ProductOfItem(entity.id, entity.name, entity.price);
-    }
